+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 9.2.1
 
+* `Core`:
+  * All `ChatMember` inheritors have fixes `status` field
+
 ## 9.2.0
 
 **Add support of [Telegram Bots API 6.9](https://core.telegram.org/bots/api-changelog#september-22-2023)**